basic scrolling

+++ src/nesCore/cpu/cpu6502.h
@@ -1,6 +1,7 @@
 #ifndef CPU6502_H_
 #define CPU6502_H_
 
+#include <cstddef>
 #include <cstdint>
 
 #define RESET_VECTOR_ADDR 0xFFFC
@@ -34,7 +35,7 @@ public:
 
     // Execute the next instruction and return
     // the number of cycle required
-    uint8_t executeInstruction(Interrupt6502 interrupt = NOINT);
+    size_t executeInstruction(Interrupt6502 interrupt = NOINT);
 
     // Return a debug struct with the current CPU status
     debug::Cpu6502Debug getDebugInfo();
